Start working on "regex labelled"
